+++ addon/components/item-picker/item-preview/component.js
@@ -21,6 +21,13 @@ export default Ember.Component.extend({
   description: Ember.computed.reads('model.description'),
 
   /**
+   * Compute the translation scope
+   */
+  _i18nScope: Ember.computed('i18nScope', function () {
+    return `${this.getWithDefault('i18nScope', 'addons.components.itemPicker')}.`;
+  }),
+
+  /**
    * What should the select button text be? we have variations depending on status
    */
   selectButtonText: Ember.computed('isValidating', 'selectAnyway', function () {