+++ tests/integration/components/item-picker/feature-service-preview/component-test.js
@@ -46,7 +46,6 @@ test('it renders', function (assert) {
   });
 
   this.render(hbs`{{item-picker/feature-service-preview
-    _i18nScope="ember-arcgis-portal-components.itemPicker."
     model=currentItem
     onSelectionValidator=onSelectionValidator
     onItemSelected=onPreviewSelected
@@ -54,7 +53,7 @@ test('it renders', function (assert) {
   }}`);
 
   const el = this.$('.item-picker-current-item-preview');
-  assert.equal(el.find('h2').text().trim(), 'This is the name');
+  assert.equal(el.find('h4').text().trim(), 'This is the name');
   assert.equal(el.find('span').text().trim(), 'Shared by: jupe');
   assert.equal(el.find('.item-picker-current-item-preview-description').text().trim(), 'This is the description');
   assert.equal(el.find('.item-picker-current-item-preview-meta > div').length, 3);